Finish USB ACM example, minimise unneeded interrupts
